+++ src/dependency-builder/types/dependency-tree-type.js
@@ -61,5 +61,6 @@ export type DependencyTreeParams = {
   filePaths: string[],
   bindingPrefix: string,
   resolveModulesConfig: ?ResolveModulesConfig,
-  visited: ?Object
+  visited: ?Object,
+  cacheProjectAst: ?Object
 };