+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- add option to isolate component into "capsule"
+
 ## [14.1.3] - 2019-06-06
 
 ### Bug fixes